image_engine_app/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-engine-app"
version = "1.0.3"
description = "Sprite Factory desktop image engine"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "Pillow>=10.0",
  "PySide6>=6.6",
]

[project.optional-dependencies]
build = [
  "pyinstaller>=6.0",
]
